chore: changed config generation
