--- .github/workflows/validate-schema.yml
@@ -1,38 +1,0 @@
-name: "Validate Schema"
-
-on:
-  workflow_dispatch:
-    inputs:
-      stateId:
-        description: "State Id"
-      eventName:
-        description: "Event Name"
-      eventPayload:
-        description: "Event Payload"
-      settings:
-        description: "Settings"
-      authToken:
-        description: "Auth Token"
-      ref:
-        description: "Ref"
-
-jobs:
-  validate:
-    name: "Validate Schema"
-    runs-on: ubuntu-latest
-    permissions: write-all
-
-    steps:
-      - uses: actions/checkout@v4
-
-      - name: Setup node
-        uses: actions/setup-node@v4
-        with:
-          node-version: "20.10.0"
-
-      - name: Install deps and run validation
-        run: |
-          yarn install --immutable --immutable-cache --check-cache
-          yarn tsx src/validate-schema.ts
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
