+++ src/helpers/validator.ts
@@ -1,5 +1,7 @@
+import * as github from "@actions/github";
+import { Octokit } from "@octokit/rest";
 import { TransformDecodeCheckError, TransformDecodeError, Value, ValueError } from "@sinclair/typebox/value";
-import { Env, envSchema, envValidator, pluginSettingsValidator, UserActivityWatcherSettings, userActivityWatcherSettingsSchema } from "../types/plugin-inputs";
+import { Env, envSchema, envValidator, pluginSettingsValidator, PluginSettings, pluginSettingsSchema } from "../types/plugin-inputs";
 
 export function validateAndDecodeSchemas(rawEnv: object, rawSettings: object) {
   const errors: ValueError[] = [];
@@ -12,7 +14,7 @@ export function validateAndDecodeSchemas(rawEnv: object, rawSettings: object) {
     }
   }
 
-  const settings = Value.Default(userActivityWatcherSettingsSchema, rawSettings) as UserActivityWatcherSettings;
+  const settings = Value.Default(pluginSettingsSchema, rawSettings) as PluginSettings;
   if (!pluginSettingsValidator.test(settings)) {
     for (const error of pluginSettingsValidator.errors(settings)) {
       console.error(error);
@@ -25,7 +27,7 @@ export function validateAndDecodeSchemas(rawEnv: object, rawSettings: object) {
   }
 
   try {
-    const decodedSettings = Value.Decode(userActivityWatcherSettingsSchema, settings);
+    const decodedSettings = Value.Decode(pluginSettingsSchema, settings);
     const decodedEnv = Value.Decode(envSchema, rawEnv || {});
     return { decodedEnv, decodedSettings };
   } catch (e) {
@@ -35,4 +37,17 @@ export function validateAndDecodeSchemas(rawEnv: object, rawSettings: object) {
     }
     throw e;
   }
+}
+
+export async function returnDataToKernel(repoToken: string, stateId: string, output: object, eventType = "return_data_to_ubiquibot_kernel") {
+  const octokit = new Octokit({ auth: repoToken });
+  return octokit.repos.createDispatchEvent({
+    owner: github.context.repo.owner,
+    repo: github.context.repo.repo,
+    event_type: eventType,
+    client_payload: {
+      state_id: stateId,
+      output: JSON.stringify(output),
+    },
+  });
 }
