+++ src/parser/payload.ts
@@ -1,12 +1,12 @@
 import * as github from "@actions/github";
-import { Value } from "@sinclair/typebox/value";
 import { config } from "dotenv";
-import { PluginInputs, userActivityWatcherSettingsSchema } from "../types/plugin-inputs";
+import { validateAndDecodeSchemas } from "../helpers/validator";
+import { PluginInputs } from "../types/plugin-inputs";
 
 config();
 
 const webhookPayload = github.context.payload.inputs;
-const settings = Value.Decode(userActivityWatcherSettingsSchema, Value.Default(userActivityWatcherSettingsSchema, JSON.parse(webhookPayload.settings)));
+const { decodedSettings } = validateAndDecodeSchemas(JSON.parse(webhookPayload.settings), process.env);
 
 const program: PluginInputs = {
   stateId: webhookPayload.stateId,
@@ -14,7 +14,7 @@ const program: PluginInputs = {
   authToken: webhookPayload.authToken,
   ref: webhookPayload.ref,
   eventPayload: JSON.parse(webhookPayload.eventPayload),
-  settings,
+  settings: decodedSettings,
 };
 
 export default program;
